+++ subscription/util/blocking/durable-subscription/src/main/java/org/occurrent/subscription/blocking/durable/DurableSubscriptionModel.java
@@ -23,8 +23,8 @@ import org.occurrent.subscription.SubscriptionPosition;
 import org.occurrent.subscription.api.blocking.*;
 
 import javax.annotation.PreDestroy;
+import java.util.Objects;
 import java.util.function.Consumer;
-import java.util.function.Supplier;
 
 import static java.util.Objects.requireNonNull;
 import static org.occurrent.subscription.PositionAwareCloudEvent.getSubscriptionPositionOrThrowIAE;
@@ -76,8 +76,27 @@ public class DurableSubscriptionModel implements PositionAwareSubscriptionModel,
     }
 
     @Override
-    public Subscription subscribe(String subscriptionId, SubscriptionFilter filter, Supplier<StartAt> startAtSupplier, Consumer<CloudEvent> action) {
-        return subscriptionModel.subscribe(subscriptionId, filter, startAtSupplier, cloudEvent -> {
+    public Subscription subscribe(String subscriptionId, SubscriptionFilter filter, StartAt startAt, Consumer<CloudEvent> action) {
+        Objects.requireNonNull(startAt, StartAt.class.getSimpleName() + " supplier cannot be null");
+
+        final StartAt startAtToUse;
+        if (startAt.isDefault()) {
+            startAtToUse = StartAt.dynamic(() -> {
+                // It's important that we find the document inside the supplier so that we lookup the latest resume token on retry
+                SubscriptionPosition subscriptionPosition = storage.read(subscriptionId);
+                if (subscriptionPosition == null) {
+                    SubscriptionPosition globalSubscriptionPosition = subscriptionModel.globalSubscriptionPosition();
+                    if (globalSubscriptionPosition != null) {
+                        subscriptionPosition = storage.save(subscriptionId, globalSubscriptionPosition);
+                    }
+                }
+                return subscriptionPosition == null ? StartAt.subscriptionModelDefault() : StartAt.subscriptionPosition(subscriptionPosition);
+            });
+        } else {
+            startAtToUse = startAt;
+        }
+
+        return subscriptionModel.subscribe(subscriptionId, filter, startAtToUse, cloudEvent -> {
                     action.accept(cloudEvent);
                     if (config.persistCloudEventPositionPredicate.test(cloudEvent)) {
                         SubscriptionPosition subscriptionPosition = getSubscriptionPositionOrThrowIAE(cloudEvent);
@@ -85,34 +104,6 @@ public class DurableSubscriptionModel implements PositionAwareSubscriptionModel,
                     }
                 }
         );
-    }
-
-    @Override
-    public Subscription subscribe(String subscriptionId, Consumer<CloudEvent> action) {
-        return subscribe(subscriptionId, (SubscriptionFilter) null, action);
-    }
-
-    /**
-     * Start listening to cloud events persisted to the event store.
-     *
-     * @param subscriptionId The id of the subscription, must be unique!
-     * @param filter         The filter to apply for this subscription. Only events matching the filter will cause the <code>action</code> to be called.
-     * @param action         This action will be invoked for each cloud event that is stored in the EventStore that matches the supplied <code>filter</code>.
-     */
-    @Override
-    public Subscription subscribe(String subscriptionId, SubscriptionFilter filter, Consumer<CloudEvent> action) {
-        Supplier<StartAt> startAtSupplier = () -> {
-            // It's important that we find the document inside the supplier so that we lookup the latest resume token on retry
-            SubscriptionPosition subscriptionPosition = storage.read(subscriptionId);
-            if (subscriptionPosition == null) {
-                SubscriptionPosition globalSubscriptionPosition = subscriptionModel.globalSubscriptionPosition();
-                if (globalSubscriptionPosition != null) {
-                    subscriptionPosition = storage.save(subscriptionId, globalSubscriptionPosition);
-                }
-            }
-            return subscriptionPosition == null ? StartAt.now() : StartAt.subscriptionPosition(subscriptionPosition);
-        };
-        return subscribe(subscriptionId, filter, startAtSupplier, action);
     }
 
     @Override